Fetch Types, N + 1 Problem, Schema Versioning, Projection

+++ src/main/java/in/yjawaid/product_service_110524/repositories/projections/ProductProjection.java
@@ -8,8 +8,8 @@ public interface ProductProjection {
 
     int getId();
     String getTitle();
-            String getDescription();
-    BigDecimal getPrice();
+    String getDescription();
+    double getPrice();
     Category getCategory();
     String getImageUrl();
 }
